requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.3
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0
